Create googledriver.py
This script takes the temperature data recorded by thermometer.py in the record.csv file and uploads it to a file on google drive called piceberg1.csv. 

This uploads the whole record.csv file as is. This can be used to "append" data to the google drive file by running it again to upload the whole file again (with the new entries since the last upload). This can be scheduled with a chron.